requirements.txt
beautifulsoup4==4.7.1
certifi==2019.3.9
chardet==3.0.4
cycler==0.10.0
dj-database-url==0.5.0
dj-static==0.0.6
Django==2.1.7
django-markdown-deux==1.0.5
django-ranged-response==0.2.0
django-simple-captcha==0.5.11
gunicorn==19.9.0
idna==2.8
kiwisolver==1.0.1
markdown2==2.3.7
matplotlib==3.0.3
numpy==1.16.2
Pillow==6.0.0
psycopg2==2.7.7
pyparsing==2.4.0
python-dateutil==2.8.0
pytz==2018.9
requests==2.21.0
selenium==3.141.0
six==1.12.0
soupsieve==1.9
static3==0.7.0
urllib3==1.24.1
